+++ packages/openapi-validation/CHANGELOG.md
@@ -1,5 +1,12 @@
+<a name="openapi-validation-2.1.0"></a>
+# [openapi-validation-2.1.0](https://github.com/ts-stack/ditsmod/releases/tag/openapi-validation-2.1.0) (2022-09-30)
+
+### Features
+
+- Added ability to override `ParametersInterceptor` and `RequestBodyInterceptor` in `providersPerReq` on module or controller level.
+
 <a name="openapi-validation-2.0.2"></a>
-# [openapi-validation-2.0.2](https://github.com/ts-stack/ditsmod/releases/tag/openapi-validation-2.0.2) (2022-09-29)
+## [openapi-validation-2.0.2](https://github.com/ts-stack/ditsmod/releases/tag/openapi-validation-2.0.2) (2022-09-29)
 
 ### Bug fix